Refactor JSON-RPC tests (#177)
