+++ test/rpc/rpc_utils.py
@@ -1,0 +1,58 @@
+"""
+Utilities for RPC tests
+"""
+
+from __future__ import annotations
+
+import json
+from typing import Union
+
+from starknet_devnet.server import app
+
+
+def rpc_call(method: str, params: Union[dict, list]) -> dict:
+    """
+    Make a call to the RPC endpoint
+    """
+    req = {
+        "jsonrpc": "2.0",
+        "method": method,
+        "params": params,
+        "id": 0
+    }
+
+    resp = app.test_client().post(
+        "/rpc",
+        content_type="application/json",
+        data=json.dumps(req)
+    )
+    result = json.loads(resp.data.decode("utf-8"))
+    return result
+
+
+def gateway_call(method: str, **kwargs):
+    """
+    Make a call to the gateway
+    """
+    resp = app.test_client().get(
+        f"/feeder_gateway/{method}?{'&'.join(f'{key}={value}&' for key, value in kwargs.items())}"
+    )
+    return json.loads(resp.data.decode("utf-8"))
+
+
+def get_block_with_transaction(transaction_hash: str) -> dict:
+    """
+    Retrieve block for given transaction
+    """
+    transaction = gateway_call("get_transaction", transactionHash=transaction_hash)
+    block_number: int = transaction["block_number"]
+    block = gateway_call("get_block", blockNumber=block_number)
+    return block
+
+
+def pad_zero(felt: str) -> str:
+    """
+    Convert felt with format `0xValue` to format `0x0Value`
+    """
+    felt = felt.lstrip("0x")
+    return "0x0" + felt
